Add JavaDoc and apply lombok annotation for all classes.
- SiteController, LinkController delete RepsonseEntity for all methods
- add new class ErrorDTO and OriginalDTO
- UrlShortcutApplication, delete RestTemplate bean
- MainService, add system of anchors and validation to all methods;
- GlobalExceptionHandler, add new private method: setResponse() and extractAnchor(); add two handlers for exception

+++ src/main/java/ru/job4j/shortcut/service/MainService.java
@@ -1,11 +1,9 @@
 package ru.job4j.shortcut.service;
 
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.shortcut.model.*;
 import ru.job4j.shortcut.repository.LinkRepository;
 import ru.job4j.shortcut.repository.SiteRepository;
@@ -16,73 +14,112 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
+/**
+ * This class contains main business logic of application
+ */
 @Service
+@AllArgsConstructor
+@Slf4j
 public class MainService {
 
+    /**
+     * DAO for site
+     */
     private final SiteRepository siteRepository;
 
+    /**
+     * DAO for link
+     */
     private final LinkRepository linkRepository;
 
+    /**
+     * Object for encoding user's password
+     */
     private BCryptPasswordEncoder encoder;
 
-    public MainService(SiteRepository siteRepository, LinkRepository linkRepository, BCryptPasswordEncoder encoder) {
-        this.siteRepository = siteRepository;
-        this.linkRepository = linkRepository;
-        this.encoder = encoder;
-    }
-
+    /**
+     * Method for getting list of all sites from database
+     * @return List of sites
+     */
     public List<Site> findAllSites() {
+        String anchor = UUID.randomUUID().toString();
+        Iterable<Site> siteList = siteRepository.findAll();
+        if (siteList == null) {
+            throw new NullPointerException("An internal error has occurred. Please try again later or contact technical support with the 'anchor'. anchor: " + anchor);
+        }
         return StreamSupport.stream(
-                this.siteRepository.findAll().spliterator(), false
+                siteList.spliterator(), false
         ).collect(Collectors.toList());
     }
 
+    /**
+     * Method for getting list of all links from database
+     * @return List of links
+     */
     public List<Link> findAllLinks() {
+        String anchor = UUID.randomUUID().toString();
+        Iterable<Link> linkList = linkRepository.findAll();
+        if (linkList == null) {
+            throw new NullPointerException("An internal error has occurred. Please try again later or contact technical support with the 'anchor'. anchor: " + anchor);
+        }
         return StreamSupport.stream(
-                this.linkRepository.findAll().spliterator(), false
+               linkList.spliterator(), false
         ).collect(Collectors.toList());
     }
 
-    public ResponseEntity<Site> findSiteById(int id) {
+    /**
+     * Method for getting site from database by ID
+     * @param id database site ID
+     * @return Object of site
+     */
+    public Site findSiteById(int id) {
+        String anchor = UUID.randomUUID().toString();
         var site = this.siteRepository.findById(id);
         if (site.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Site not found");
+            throw new IllegalArgumentException("Site not found. Actual parameters: site ID - " + id + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
-        return ResponseEntity.status(HttpStatus.OK)
-                .header("Content-Type", "application/json")
-                .body(site.get());
+        return site.get();
     }
 
-    public ResponseEntity<Link> findLinkById(int id) {
+    /**
+     * Method for getting link from database by link ID
+     * @param id - database link ID
+     * @return Link object
+     */
+    public Link findLinkById(int id) {
+        String anchor = UUID.randomUUID().toString();
         var link = this.linkRepository.findById(id);
         if (link.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Link not found");
+            throw new IllegalArgumentException("Link not found. Actual parameters: link ID - " + id + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
-        return ResponseEntity.status(HttpStatus.OK)
-                .header("Content-Type", "application/json")
-                .body(link.get());
+        return link.get();
     }
 
-    public ResponseEntity<SiteDTO> saveSite(SiteOnlyNameDTO siteOnlyNameDTO) {
+    /**
+     * Method for saving new site in database
+     * @param siteOnlyNameDTO - DTO model containing: name of site
+     * @return SiteDTO model that contained: registered status, login and password
+     */
+    public SiteDTO saveSite(SiteOnlyNameDTO siteOnlyNameDTO) {
         Site check = siteRepository.findByNameOfSite(siteOnlyNameDTO.getNameOfSite());
         if (check != null) {
-            return new ResponseEntity<>(
-                    new SiteDTO(true, check.getUsername(), "hidden password"),
-                    HttpStatus.OK
-            );
+            return new SiteDTO(true, check.getUsername(), "hidden password");
         } else {
             Site buff = Site.of(siteOnlyNameDTO.getNameOfSite());
             String passBefore = buff.getPassword();
             buff.setPassword(encoder.encode(passBefore));
             Site siteAfterSave = this.siteRepository.save(buff);
-            return new ResponseEntity<>(
-                    new SiteDTO(false, siteAfterSave.getUsername(), passBefore),
-                    HttpStatus.CREATED
-            );
+            return new SiteDTO(false, siteAfterSave.getUsername(), passBefore);
         }
     }
 
-    public ResponseEntity<UrlShortcutDTO> saveLink(LinkDTO linkDTO) {
+    /**
+     * Method for saving short link to database
+     * @param linkDTO - DTO model containing: site ID and initial URL
+     * @return UrlShortcutDTO model contaning: string with shortened link
+     */
+    public UrlShortcutDTO saveLink(LinkDTO linkDTO) {
+        String anchor = UUID.randomUUID().toString();
         Optional<Site> buffSite = siteRepository.findById(linkDTO.getSiteId());
         String rootUrlFromLinkDTO;
         String linkForSave;
@@ -95,19 +132,23 @@ public class MainService {
         }
         if (buffSite.isPresent() && buffSite.get().getNameOfSite().equals(rootUrlFromLinkDTO)) {
             Link buff = this.linkRepository.save(Link.of(linkForSave, buffSite.get()));
-            UrlShortcutDTO urlShortcutDTO = new UrlShortcutDTO("http://localhost:8080/link/redirect/" + buff.getShortcut());
-            return new ResponseEntity<>(
-                    urlShortcutDTO,
-                    HttpStatus.CREATED);
+            return new UrlShortcutDTO("http://localhost:8080/link/redirect/" + buff.getShortcut());
         } else {
-            throw new NullPointerException("Site not registered, please register");
+            throw new IllegalArgumentException("Site not found. Actual parameters: site ID - " + linkDTO.getSiteId() + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
     }
 
-    public ResponseEntity<Void> updateSite(Site site) throws InvocationTargetException, IllegalAccessException {
+    /**
+     * Method for updating information about of site (ID, name of site, username, password)
+     * @param site - object with information about of site
+     * @throws InvocationTargetException
+     * @throws IllegalAccessException
+     */
+    public void updateSite(Site site) throws InvocationTargetException, IllegalAccessException {
+        String anchor = UUID.randomUUID().toString();
         var current = siteRepository.findById(site.getId());
         if (current.isEmpty()) {
-            throw new NullPointerException("Site with this id not found");
+            throw new IllegalArgumentException("Site not found. Actual parameters: site ID - " + site.getId() + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
         var buffSite = current.get();
         var methods = buffSite.getClass().getDeclaredMethods();
@@ -123,7 +164,7 @@ public class MainService {
                 var getMethod = namePerMethod.get(name);
                 var setMethod = namePerMethod.get(name.replace("get", "set"));
                 if (setMethod == null) {
-                    throw new NullPointerException("Invalid properties");
+                    throw new NullPointerException("An internal error has occurred. Please try again later or contact technical support with the 'anchor'. anchor: " + anchor);
                 }
                 var newValue = getMethod.invoke(site);
                 if (newValue != null) {
@@ -133,13 +174,19 @@ public class MainService {
         }
         buffSite.setPassword(encoder.encode(buffSite.getPassword()));
         siteRepository.save(buffSite);
-        return ResponseEntity.ok().build();
     }
 
-    public ResponseEntity<Void> updateLink(Link link) throws InvocationTargetException, IllegalAccessException {
+    /**
+     * Method for updating link object
+     * @param link - Link object
+     * @throws InvocationTargetException
+     * @throws IllegalAccessException
+     */
+    public void updateLink(Link link) throws InvocationTargetException, IllegalAccessException {
+        String anchor = UUID.randomUUID().toString();
         var current = linkRepository.findById(link.getId());
         if (current.isEmpty()) {
-            throw new NullPointerException("Link not found");
+            throw new IllegalArgumentException("Link not found. Actual parameters: link ID - " + link.getId() + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
         var buffLink = current.get();
         var methods = buffLink.getClass().getDeclaredMethods();
@@ -155,7 +202,7 @@ public class MainService {
                 var getMethod = namePerMethod.get(name);
                 var setMethod = namePerMethod.get(name.replace("get", "set"));
                 if (setMethod == null) {
-                    throw new NullPointerException("Invalid properties");
+                    throw new NullPointerException("An internal error has occurred. Please try again later or contact technical support with the 'anchor'. anchor: " + anchor);
                 }
                 var newValue = getMethod.invoke(link);
                 if (newValue != null) {
@@ -164,53 +211,72 @@ public class MainService {
             }
         }
         linkRepository.save(buffLink);
-        return ResponseEntity.ok().build();
     }
 
-    public ResponseEntity<Void> deleteSite(int id) {
+    /**
+     * Method for deleting site from database by site ID
+     * @param id - database site ID
+     */
+    public void deleteSite(int id) {
+        String anchor = UUID.randomUUID().toString();
         if (this.siteRepository.findById(id).isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Site not found");
+            throw new IllegalArgumentException("Site not found. Actual parameters: site ID - " + id + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
-        Site site = new Site();
-        site.setId(id);
-        this.siteRepository.delete(site);
-        return ResponseEntity.ok().build();
+        if (!linkRepository.findBySiteId(id).isEmpty()) {
+            throw new IllegalArgumentException("Site cannot be deleted because it contains active links. Actual parameters: site ID - " + id + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
+        } else {
+            Site site = new Site();
+            site.setId(id);
+            this.siteRepository.delete(site);
+        }
     }
 
-    public ResponseEntity<Void> deleteLink(int id) {
+    /**
+     * Method for deleting link by link ID
+     * @param id - link ID
+     */
+    public void deleteLink(int id) {
+        String anchor = UUID.randomUUID().toString();
         if (this.linkRepository.findById(id).isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Link not found");
+            throw new IllegalArgumentException("Link not found. Actual parameters: link ID - " + id + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
         Link link = new Link();
         link.setId(id);
         this.linkRepository.delete(link);
-        return ResponseEntity.ok().build();
     }
 
-    public ResponseEntity<Void> redirectLink(String shortUrl) {
+    /**
+     * Method for getting original link by shorted identifier
+     * @param shortUrl - shorted identifier
+     * @return OriginalUrlDTO object
+     */
+    public OriginalUrlDTO redirectLink(String shortUrl) {
+        String anchor = UUID.randomUUID().toString();
         var link = this.linkRepository.findByShortcut(shortUrl);
         if (link == null) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Shortcut not found");
+            throw new IllegalArgumentException("Short url not found. Actual parameters: shortUrl - " + shortUrl + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
         }
         this.linkRepository.incrementTotal();
-        HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.add("Location", "http://" + link.getUrl());
-        return new ResponseEntity<>(
-                httpHeaders, HttpStatus.MOVED_PERMANENTLY
-        );
+        return new OriginalUrlDTO(link.getUrl());
     }
 
-    public ResponseEntity<StatLinkListDTO> statLinkForSite(int siteId) {
+    /**
+     * Method for getting shortened link call statistics for the site by site ID
+     * @param siteId - site ID
+     * @return StatLinkListDTO object
+     */
+    public StatLinkListDTO statLinkForSite(int siteId) {
+        String anchor = UUID.randomUUID().toString();
         List<StatLinkDTO> buffListDTO = new ArrayList<>();
-        Collection<Link> buffLinkList = StreamSupport.stream(
-                this.linkRepository.findBySiteId(siteId).spliterator(), false
-        ).collect(Collectors.toList());
-        for (Link elem : buffLinkList) {
+        Collection<Link> linkCollection = linkRepository.findBySiteId(siteId);
+        if (linkCollection == null) {
+            throw new NullPointerException("An internal error has occurred. Please try again later or contact technical support with the 'anchor'. anchor: " + anchor);
+        } else if (linkCollection.isEmpty()) {
+            throw new IllegalArgumentException("Site not found. Actual parameters: site ID - " + siteId + ". Please contact technical support with the 'anchor'. anchor: " + anchor);
+        }
+        for (Link elem : linkCollection) {
             buffListDTO.add(new StatLinkDTO(elem.getUrl(), elem.getTotal()));
         }
-        StatLinkListDTO statLinkListDTO = new StatLinkListDTO(buffListDTO);
-        return new ResponseEntity<>(
-                statLinkListDTO,
-                HttpStatus.OK);
+        return new StatLinkListDTO(buffListDTO);
     }
 }
